ENH: Expose superagent module as an input parameter and ECMA6 rewrite
Exposing the superagent module to the user facilitates
use of superagent-cache module for fetching requests.
This can be either an in-memory cache or a redis instance.
This is useful because of the hard Google Books API request
limit of 1000 reqeusts/day and decreases round trip time
on cache hit.

This commit also turns the module into an ECMA6 class and
wraps the search function in a promise.